java-note | Update Format
