--- JavaCore/src/main/java/config/JavaConfig.java
@@ -1,11 +1,0 @@
-package config;
-
-import com.github.javafaker.Faker;
-
-public class JavaConfig {
-    public static Faker faker;
-
-    static {
-        faker = new Faker();
-    }
-}
