+++ JavaCore/src/main/java/encapsulation_inheritance_polymorphism/Encapsulation & Inheritance & Polymorphism.md
@@ -1,0 +1,1 @@
+## Encapsulation & Inheritance & Polymorphism
