+++ JavaCore/src/main/java/fundamentals/Basic Syntax And Basic OOP.md
