+++ JavaCore/src/main/java/fundamentals/BestPracticeForCondition.md
