--- JavaCore/src/main/java/java8/inteface/App.java
@@ -1,15 +1,0 @@
-package java8.inteface;
-
-public class App {
-    public static void main(String[] args) {
-        DefaultAndStaticInterface overrideDefaultMethod = new DefaultAndStaticInterface1Impl();
-        DefaultAndStaticInterface noOverrideDefaultMethod = new DefaultAndStaticInterface2Impl();
-
-
-        overrideDefaultMethod.defaultMethod();
-        //==> Default Method implemented Custom (overridden)
-
-        noOverrideDefaultMethod.defaultMethod();
-        //==> defaultMethod (still Default)
-    }
-}
