--- JavaCore/src/main/java/java8/inteface/DefaultAndStaticInterface2Impl.java
@@ -1,8 +1,0 @@
-package java8.inteface;
-
-public class DefaultAndStaticInterface2Impl implements DefaultAndStaticInterface {
-    @Override
-    public String handleFirstName(String string) {
-        return "Handle First Name";
-    }
-}
