--- JavaCore/src/main/java/java9/ModuleSystem.java
@@ -1,5 +1,0 @@
-package java9;
-
-public class ModuleSystem {
-    //https://www.digitalocean.com/community/tutorials/java-9-features-with-examples#java-9-module-system
-}
