--- JavaCore/src/main/java/java9/inteface/App.java
@@ -1,9 +1,0 @@
-package java9.inteface;
-
-public class App {
-
-    public static void main(String[] args) {
-        PrivateMethodInterface privateMethodInterface = new PrivateMethodInterfaceImpl();
-        System.out.println(privateMethodInterface.getCardID());
-    }
-}
