--- JavaCore/src/main/java/java9/inteface/PrivateMethodInterface.java
@@ -1,17 +1,0 @@
-package java9.inteface;
-
-public interface PrivateMethodInterface {
-    private Long createCardID(){
-        // Method implementation goes here.
-        System.out.println("Create Card ID");
-        return 0L;
-    }
-
-    default Long getCardID(){
-        return createCardID();
-    }
-
-    private static void displayCardDetails(){
-        // Method implementation goes here.
-    }
-}
