--- JavaCore/src/main/java/java9/inteface/PrivateMethodInterfaceImpl.java
@@ -1,5 +1,0 @@
-package java9.inteface;
-
-public class PrivateMethodInterfaceImpl implements PrivateMethodInterface{
-
-}
