+++ JavaCore/src/main/java/old_package/config/JavaConfig.java
@@ -1,0 +1,11 @@
+package old_package.config;
+
+import com.github.javafaker.Faker;
+
+public class JavaConfig {
+    public static Faker faker;
+
+    static {
+        faker = new Faker();
+    }
+}
