+++ JavaCore/src/main/java/old_package/java8/inteface/DefaultAndStaticInterface2Impl.java
@@ -1,0 +1,8 @@
+package old_package.java8.inteface;
+
+public class DefaultAndStaticInterface2Impl implements DefaultAndStaticInterface {
+    @Override
+    public String handleFirstName(String string) {
+        return "Handle First Name";
+    }
+}
