+++ JavaCore/src/main/java/old_package/java9/ModuleSystem.java
@@ -1,0 +1,5 @@
+package old_package.java9;
+
+public class ModuleSystem {
+    //https://www.digitalocean.com/community/tutorials/java-9-features-with-examples#java-9-module-system
+}
