+++ JavaCore/src/main/java/old_package/java9/inteface/App.java
@@ -1,0 +1,9 @@
+package old_package.java9.inteface;
+
+public class App {
+
+    public static void main(String[] args) {
+        PrivateMethodInterface privateMethodInterface = new PrivateMethodInterfaceImpl();
+        System.out.println(privateMethodInterface.getCardID());
+    }
+}
