+++ JavaCore/src/main/java/old_package/java9/inteface/PrivateMethodInterfaceImpl.java
@@ -1,0 +1,5 @@
+package old_package.java9.inteface;
+
+public class PrivateMethodInterfaceImpl implements PrivateMethodInterface{
+
+}
