+++ JavaCore/src/main/java/strings/Strings.md
@@ -1,0 +1,1 @@
+## Strings
